validacion en html y js de los comentarios y likes

+++ routes/comentario.py
@@ -3,18 +3,18 @@ from configuration import db
 from models.comentario import Comentario
 from . import routes
 import http
+from datetime import datetime
 
-
-@routes.route("/comentario/")
-def getComentarios():
+@routes.route("/comentario/<int:movieID>")
+def getComentarios(movieID):
     status_code=0
     res = Comentario.query.all()
     for i in range(0,len(res)):
-        print(Comentario.query.filter_by(id=res[i].id).first().ilikes)
         res[i].data["likes"]=len(Comentario.query.filter_by(id=res[i].id).first().ilikes)
     sorted(res, key=lambda x: x.data["likes"], reverse=True)
     status_code =http.HTTPStatus.ACCEPTED
-    return jsonify({"cometarios": [m.toJson() for m in res],"status_code":status_code})	
+    return jsonify({"comentarios": [m.toJson() for m in res if m.data['idMovie']== movieID],"status_code":status_code})	
+
 
 
 @routes.route("/comentario/", methods=["POST"])
@@ -30,7 +30,8 @@ def createComentario():
         resDATA={
             data1[0][0][1:len(data1[0][0])-1]:data1[0][1][1:len(data1[0][1])-1],
             data1[1][0][1:len(data1[1][0])-1]:int(data1[1][1]),
-            data1[2][0][1:len(data1[2][0])-1]:int(data1[2][1])
+            data1[2][0][1:len(data1[2][0])-1]:int(data1[2][1]),
+            data1[3][0][1:len(data1[3][0])-1]:datetime.now().strftime("%d/%m/%Y %H:%M:%S"),
         }
         comentario = Comentario(data=resDATA)
         db.session.add(comentario)
@@ -65,21 +66,26 @@ def updateComentario(commentID):
         db.session.close()
     return jsonify({"message":message,'status_code':status_code})
 
-@routes.route("/comentario/<int:commentID>", methods=["DELETE"])
-def deleteComentario(commentID):
+@routes.route("/comentario/<int:commentID>/<int:userID>", methods=["DELETE"])
+def deleteComentario(commentID,userID):
     message = ""
+    error = False
     status_code=0
     try:
         comentario = Comentario.query.filter_by(id=commentID).first()
-        db.session.delete(comentario)
-        db.session.commit()
-        message = "comentario eliminada con exito"
-        status_code = http.HTTPStatus.ACCEPTED
+        if comentario.data['idUser']==userID:
+            db.session.delete(comentario)
+            db.session.commit()
+            message = "comentario eliminada con exito"
+            status_code = http.HTTPStatus.ACCEPTED
+        else:
+            message = "usuario no autorizado"
     except Exception as e:
         db.session.rollback()
+        error=True
         status_code = http.HTTPStatus.INTERNAL_SERVER_ERROR
-        abort(status_code)
     finally:
         db.session.close()
-
+    if error == True:
+        abort(status_code)
     return jsonify({"message": message,'status_code':status_code})
